Fixed int overflow when specifiying upload size

+++ src/main/scala/org/mbari/m3/panoptes/api/ImageV1Api.scala
@@ -37,7 +37,7 @@ class ImageV1Api(maxFileSizeGB: Int = 3)(implicit val executor: ExecutionContext
 
   // Configure Max upload size
   configureMultipartHandling(
-    MultipartConfig(maxFileSize = Some(maxFileSizeGB * 1024 * 1024 * 1024))
+    MultipartConfig(maxFileSize = Some(maxFileSizeGB.toLong * 1024L * 1024L * 1024L))
   )
 
   post("/:camera_id/:deployment_id/:name") {
